requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
dj-email-url==1.0.6
dj-rest-auth==2.1.12
Django==4.0.10
django-allauth==0.48.0
django-cache-url==3.4.4
django-cors-headers==3.10.1
djangorestframework==3.13.1
drf-spectacular==0.21.2
environs==9.5.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jsonschema==4.18.2
jsonschema-specifications==2023.6.1
marshmallow==3.19.0
oauthlib==3.2.2
packaging==23.1
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.8.10
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
whitenoise==5.3.0
